+++ backend/xchange_api/src/main/java/br/com/xchange/api/domain/entities/CoinWithMarketData.java
@@ -1,12 +1,14 @@
 package br.com.xchange.api.domain.entities;
 
+import java.math.BigDecimal;
+
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 
 @Data
 @EqualsAndHashCode(callSuper=false)
 public class CoinWithMarketData extends Coin {
-  private Double priceBtc;
+  private BigDecimal priceBtc;
   private String marketCap;
   private String totalVolume;
   private String sparkline;
